Move LiftedOperators to pretty-tests.
